+++ cypress/locators/vertical-menu/locators.js
@@ -1,0 +1,6 @@
+// component preview locators
+export const VERTICAL_MENU_COMPONENT = '[data-component="vertical-menu"]';
+export const VERTICAL_MENU_ITEM = '[data-component="vertical-menu-item"]';
+export const VERTICAL_MENU_TRIGGER = '[data-component="vertical-menu-trigger"]';
+export const VERTICAL_MENU_FULL_SCREEN =
+  '[data-component="vertical-menu-full-screen"]';